Fixes the map and a compile error
I made a doof when I changed the map last in the robotics maint area
while trying to figure out why that door wasn't working before (It turns
out it was in a 'space' area, so started unpowered.). Fixed the
scrubber/disposal pipe and the wire that was supposed to be there.

Also noticed the last commit had a few missing quotations that prevented
compiling, so I fixed that too.